Extract out new module `TokenPolicyId.Gen`.
